healthcare_project/requirements.txt
# Required dependencies 
Flask
Flask-CORS
google-generativeai
